// apps/ui/src/app/client/model/chatMcpOverrideDto.ts
/**
 * LMStudio Wrapper API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ChatMcpOverrideDto { 
    /**
     * References User.customMcps[].id
     */
    mcpId: string;
    active: boolean;
    allowedTools: Array<string>;
}