// apps/ui/src/app/client/model/customMcpDto.ts
/**
 * LMStudio Wrapper API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CustomMcpDto { 
    id: string;
    name: string;
    endpoint: string;
    active: boolean;
    availableTools: Array<string>;
    allowedTools: Array<string>;
    /**
     * Custom HTTP headers sent to this MCP server
     */
    headers?: object;
}